+++ src/pages/ReviewListPage/ReviewListPage.tsx
@@ -19,7 +19,6 @@ import {
   FEASIBLE_YES,
   CONFLICT_REASONS
 } from '@utils/constants.ts';
-import GetPanelList from '@services/axios/get/getPanelList/getPanelList';
 import GetProposalByStatusList from '@services/axios/get/getProposalByStatusList/getProposalByStatusList';
 import ScienceIcon from '../../components/icon/scienceIcon/scienceIcon';
 import Alert from '../../components/alerts/standardAlert/StandardAlert';
@@ -29,7 +28,6 @@ import { PMT } from '@/utils/constants';
 import SubmitButton from '@/components/button/Submit/Submit';
 import { ProposalReview, ScienceReview, TechnicalReview } from '@/utils/types/proposalReview';
 import SubmitIcon from '@/components/icon/submitIcon/submitIcon';
-import { Panel } from '@/utils/types/panel';
 import TechnicalIcon from '@/components/icon/technicalIcon/technicalIcon';
 import PageFooterPMT from '@/components/layout/pageFooterPMT/PageFooterPMT';
 import PutProposalReview from '@/services/axios/put/putProposalReview/putProposalReview';
@@ -65,7 +63,6 @@ export default function ReviewListPage() {
   const [filteredData, setFilteredData] = React.useState<FilteredItem[]>([]);
 
   const [reset, setReset] = React.useState(false);
-  const [panelData, setPanelData] = React.useState<Panel[]>([]);
   const [proposals, setProposals] = React.useState<Proposal[]>([]);
   const [proposalReviews, setProposalReviews] = React.useState<ProposalReview[]>([]);
 
@@ -78,18 +75,6 @@ export default function ReviewListPage() {
   React.useEffect(() => {
     setReset(!reset);
   }, []);
-
-  React.useEffect(() => {
-    const GetReviewPanels = async () => {
-      const response = await GetPanelList(authClient);
-      if (typeof response === 'string') {
-        notifyError(response);
-      } else {
-        setPanelData((response as unknown) as Panel[]);
-      }
-    };
-    GetReviewPanels();
-  }, [reset]);
 
   React.useEffect(() => {
     const fetchProposalReviewData = async (_proposalId: string) => {
@@ -110,18 +95,12 @@ export default function ReviewListPage() {
       if (typeof response === 'string') {
         notifyError(response);
       } else {
-        const panelProposalIds = panelData.flatMap(panel =>
-          Array.isArray(panel.proposals) ? panel.proposals.map(proposal => proposal.proposalId) : []
-        );
-        const filtered = response
-          ? response.filter((proposal: Proposal) => panelProposalIds.includes(proposal.id))
-          : [];
-        setProposals(filtered);
-        loopProposals(filtered);
+        setProposals(response);
+        loopProposals(response);
       }
     };
     fetchProposalData();
-  }, [panelData]);
+  }, [reset]);
 
   React.useEffect(() => {
     const data = proposals ? filterProposals() : [];
@@ -438,18 +417,10 @@ export default function ReviewListPage() {
     headerName: t('dateAssigned.label'),
     width: 180,
     renderCell: (e: { row: any }) => {
-      const panel = panelData.find(
-        panel =>
-          Array.isArray(panel.proposals) &&
-          panel.proposals.some(p => p.proposalId === e.row.proposal?.id)
-      );
-      let proposal = null;
-      if (panel && panel.proposals && panel.proposals.length > 0) {
-        proposal = panel.proposals.find(p => p.proposalId === e.row.proposal?.id);
-      }
-      return proposal && proposal.assignedOn
-        ? presentDate(proposal.assignedOn) + ' ' + presentTime(proposal.assignedOn)
-        : '';
+      // TODO retrieve assigned_on from reviewable response once backend updated & update type + mapping
+      return e.row?.proposal?.assignedOn
+        ? presentDate(e.row.proposal?.assignedOn)
+        : t('unavailable');
     }
   };