+++ src/services/axios/post/postProposal/postProposal.test.tsx
@@ -27,9 +27,9 @@ describe('Helper Functions', () => {
     const proposalBackEnd: ProposalBackend = mappingPostProposal(proposal, PROPOSAL_STATUS.DRAFT);
     expect(proposalBackEnd).to.deep.equal({
       ...MockProposalBackend,
-      info: {
-        ...MockProposalBackend.info,
-        proposal_type: { ...MockProposalBackend.info.proposal_type, attributes: [] }
+      proposal_info: {
+        ...MockProposalBackend.proposal_info,
+        proposal_type: { ...MockProposalBackend.proposal_info.proposal_type, attributes: [] }
       }
     });
   });